remove deprecations (#195)
* remove deprecations

* allow failures on nightly

* add appveyor build

* test cleanup: Int64 -> Int

+++ test/test_collect.jl
@@ -111,7 +111,7 @@ end
 
     v = ((a=i,) => (b="a$i",) for i in 1:3)
     @test collect_columns(v) == Columns(Columns((a = [1,2,3],))=>Columns((b = ["a1","a2","a3"],)))
-    @test eltype(collect_columns(v)) == Pair{NamedTuple{(:a,), Tuple{Int64}}, NamedTuple{(:b,), Tuple{String}}}
+    @test eltype(collect_columns(v)) == Pair{NamedTuple{(:a,), Tuple{Int}}, NamedTuple{(:b,), Tuple{String}}}
 
     v = (i == 1 ? (a="1",) => (b="a$i",) : (a=i,) => (b="a$i",) for i in 1:3)
     @test collect_columns(v) == Columns(Columns((a = ["1",2,3],))=>Columns((b = ["a1","a2","a3"],)))
